+++ ads/aqua/extension/utils.py
@@ -1,12 +1,15 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 # Copyright (c) 2024 Oracle and/or its affiliates.
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
 from dataclasses import fields
+from datetime import datetime, timedelta
 from typing import Dict, Optional
 
+from cachetools import TTLCache, cached
 from tornado.web import HTTPError
 
+from ads.aqua import ODSC_MODEL_COMPARTMENT_OCID
+from ads.aqua.common.utils import fetch_service_compartment
 from ads.aqua.extension.errors import Errors
 
 
@@ -21,3 +24,11 @@ def validate_function_parameters(data_class, input_data: Dict):
             raise HTTPError(
                 400, Errors.MISSING_REQUIRED_PARAMETER.format(required_parameter)
             )
+
+
+@cached(cache=TTLCache(maxsize=1, ttl=timedelta(minutes=1), timer=datetime.now))
+def ui_compatability_check():
+    """This method caches the service compartment OCID details that is set by either the environment variable or if
+    fetched from the configuration. The cached result is returned when multiple calls are made in quick succession
+    from the UI to avoid multiple config file loads."""
+    return ODSC_MODEL_COMPARTMENT_OCID or fetch_service_compartment()